Moved color change from its temporary placement
